+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.3.0] - 2021-01-28
 [0.3.0]: https://github.com/mhassan1/yarn-plugin-licenses/compare/v0.2.1...v0.3.0
 
+- Refactor `getPackageManifest` into `getPackagePath` in linkers
 - Compress paths in `ppath.join`
 
 ## [0.2.1] - 2020-12-30